refactor: move state and receipt_watcher from manager
